requirements.txt
requests
beautifulsoup4
pyyaml
sentence-transformers
faiss-cpu
numpy
pandas
streamlit
flask
tqdm
huggingface_hub